:recycle: refactor: separação de escopos de utilidades
