+++ grails-app/services/mini/asaas/CustomerService.groovy
@@ -3,7 +3,7 @@ package mini.asaas
 import mini.asaas.adapters.SaveCustomerAdapter
 import mini.asaas.utils.CpfCnpjUtils
 import mini.asaas.utils.DomainUtils
-import mini.asaas.utils.Utils
+import mini.asaas.utils.EmailUtils
 
 import grails.gorm.transactions.Transactional
 import grails.validation.ValidationException
@@ -11,12 +11,13 @@ import grails.validation.ValidationException
 @Transactional
 class CustomerService {
 
+    AddressService addressService
+
     public Customer create(SaveCustomerAdapter saveCustomerAdapter) {
 
         Customer customer = validate(saveCustomerAdapter)
-        if (customer.hasErrors()) {
-            throw new ValidationException("Erro ao criar conta do usuário", customer.errors)
-        }
+
+        if (customer.hasErrors()) throw new ValidationException("Erro ao criar conta do usuário", customer.errors)
 
         Address address = new Address(
             address: saveCustomerAdapter.address.address,
@@ -46,12 +47,12 @@ class CustomerService {
 
         if (!saveCustomerAdapter.email) DomainUtils.addError(customer, "O email é obrigatório")
 
-        if (saveCustomerAdapter.email && !Utils.emailIsValid(saveCustomerAdapter.email)) DomainUtils.addError(customer, "O email informado é inválido")
+        if (saveCustomerAdapter.email && !EmailUtils.isValid(saveCustomerAdapter.email)) DomainUtils.addError(customer, "O email informado é inválido")
 
         if (!saveCustomerAdapter.cpfCnpj) DomainUtils.addError(customer, "O CPF/CNPJ é obrigatório")
 
         if (saveCustomerAdapter.cpfCnpj && !CpfCnpjUtils.validate(saveCustomerAdapter.cpfCnpj)) DomainUtils.addError(customer, "O CPF/CNPJ informado não é válido")
 
-        return Utils.validateAddress(saveCustomerAdapter.address, customer)
+        return addressService.validate(saveCustomerAdapter.address, customer)
     }
 }
