+++ src/main/groovy/mini/asaas/utils/StringUtils.groovy
@@ -1,0 +1,10 @@
+package mini.asaas.utils
+
+class StringUtils {
+
+    public static String removeNonNumeric(String text) {
+        if (text == null) return null
+
+        return text?.replaceAll("\\D+","")
+    }
+}
